Anexo Product CRUD

+++ app/Repositories/Eloquent/EloquentUserRepository.php
@@ -5,29 +5,27 @@ namespace App\Repositories\Eloquent;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Facades\Hash;
-use App\Repositories\Contracts\Users\UsersRepositoryInterface;
+use App\Repositories\Contracts\UserRepositoryInterface;
 use App\Http\DTOs\AuthUsersDTO;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
-class EloquentUsersRepository implements UsersRepositoryInterface {
+class EloquentUserRepository implements UserRepositoryInterface {
 
-    public function create(AuthUsersDTO $authUsersDTO): void {
+    public function store(AuthUsersDTO $authUsersDTO): void {
         
         $passwordHash = Hash::make($authUsersDTO->getPassword());
         
         User::create([
-            'name'=> $authUsersDTO->getName(),
-            'email'=> $authUsersDTO->getEmail(),
-            'password'=> $passwordHash,
+            'name' => $authUsersDTO->getName(),
+            'email' => $authUsersDTO->getEmail(),
+            'password' => $passwordHash,
         ]);
     }
-
     public function list(): Collection {
         
         $users = User::all();
-
-        return $users;
+        return $users;        
     }
-
     public function update(AuthUsersDTO $authUsersDTO): void {
         
         $user = User::find($authUsersDTO->getUserId());
@@ -40,12 +38,15 @@ class EloquentUsersRepository implements UsersRepositoryInterface {
         
         $user->save();
     }
-
     public function delete($userId): void {
-        
+
         $user = User::find($userId);
         
-        $user->delete();
+        if($user){
+            $user->delete();
+        }else{
+        throw new ModelNotFoundException("User not found with ID: $userId.");
+        }
     }
 
     public function findByEmail(string $email):?user
